chore: nuke dead routes and reorganize sidebar nav
Cleans up orphaned TSX routes that weren't reachable via the sidebar
and reorganizes the remaining keeps into idiomatic parent/child groups.

Removed routes:
- /card-database (duplicate of /gatherer, which is the canonical item
  card browser; if the card-database variant is ever wanted back it
  should be rebuilt in-idiom rather than resurrected)
- /japanpicture (dead)
- /press (dead)

Sidebar reorganization under Duchamp:
- /rhonda-shearer-archive nested under /rhonda-shearer alongside
  The Impossible Bed
- /readymades-research nested under /readymades
- /chess-research nested under /chess
- /duchamp-works and /la-chose-en-soie added as top-level items

Critiques section:
- New 'Critiques Index' link prepended so the index page is reachable
  without typing the URL

404 page:
- Moved app/not-found.tsx into app/(site)/not-found.tsx so it inherits
  the site layout and the sidebar. The standalone CTA buttons are
  gone; the sidebar is navigation enough.

Sitemap:
- Dropped /card-database from corePages
- Added /duchamp-works to evidencePages since it is kept

Temporary 'Orphaned (audit)' sidebar section removed now that every
orphan has been dispositioned.
